refactored build arguments inventory functions, alias, and history functions; added filenames to arginv & set general function to append HOME dir to filename
