+++ free.c
@@ -18,6 +18,7 @@ int freeall(arg_inventory_t *arginv)
 		free(arginv->history_file);
 		free_environ(arginv->envlist);
 		free_alias(arginv->alias);
+		free(arginv->alias_file);
 		if (arginv->input_commands)
 			free(arginv->input_commands);
 		exit_status = arginv->exit_status;
